--- apollo-runtime-kotlin/src/commonMain/kotlin/com/apollographql/apollo3/network/HttpMethod.kt
@@ -1,5 +1,0 @@
-package com.apollographql.apollo3.network
-
-enum class HttpMethod {
-  Get, Post
-}
