+++ apollo-runtime-kotlin/src/commonMain/kotlin/com/apollographql/apollo3/network/http/ApolloHttpNetworkTransport.kt
@@ -1,12 +1,25 @@
 package com.apollographql.apollo3.network.http
 
-import com.apollographql.apollo3.network.HttpMethod
+import com.apollographql.apollo3.ApolloRequest
+import com.apollographql.apollo3.api.Operation
+import com.apollographql.apollo3.api.Response
+import com.apollographql.apollo3.api.ResponseAdapterCache
+import com.apollographql.apollo3.api.fromResponse
+import com.apollographql.apollo3.api.internal.OperationRequestBodyComposer
+import com.apollographql.apollo3.api.variablesJson
+import com.apollographql.apollo3.exception.ApolloHttpException
+import com.apollographql.apollo3.exception.ApolloSerializationException
 import com.apollographql.apollo3.network.NetworkTransport
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.flow
+import okio.Buffer
+import okio.BufferedSink
+import okio.ByteString
 
-expect class ApolloHttpNetworkTransport constructor(
-    serverUrl: String,
-    headers: Map<String, String>,
-    httpMethod: HttpMethod = HttpMethod.Post,
+class ApolloHttpNetworkTransport(
+    val serverUrl: String,
+    val httpMethod: HttpMethod = HttpMethod.Post,
+
     /**
      * The timeout interval to use when connecting
      *
@@ -21,4 +34,97 @@ expect class ApolloHttpNetworkTransport constructor(
      * - on Android, it is used to set  [OkHttpClient.readTimeout]
      */
     readTimeoutMillis: Long = 60_000,
-) : NetworkTransport
+) : NetworkTransport {
+    val engine = HttpEngine(connectTimeoutMillis, readTimeoutMillis)
+    override fun <D : Operation.Data> execute(request: ApolloRequest<D>, responseAdapterCache: ResponseAdapterCache): Flow<Response<D>> {
+        val httpRequest = request.toHttpRequest(responseAdapterCache)
+        return flow {
+            emit(engine.execute(httpRequest) {
+                it.parse(request, responseAdapterCache)
+            })
+        }
+    }
+
+    private fun <D : Operation.Data> HttpResponse.parse(
+        request: ApolloRequest<D>,
+        responseAdapterCache: ResponseAdapterCache
+    ): Response<D> {
+        if (statusCode !in 200..299) {
+            throw ApolloHttpException(
+                statusCode = statusCode,
+                headers = headers,
+                message = "Http request failed with status code `${statusCode} (${body.readUtf8()})`"
+            )
+        }
+
+        return request.operation.fromResponse(
+            source = body,
+            responseAdapterCache = responseAdapterCache
+        ).copy(
+            requestUuid = request.requestUuid,
+            executionContext = request.executionContext + HttpResponseInfo(
+                statusCode = statusCode,
+                headers = headers
+            )
+        )
+    }
+
+    private fun <D : Operation.Data> ApolloRequest<D>.toHttpRequest(
+        responseAdapterCache: ResponseAdapterCache
+    ): HttpRequest {
+        try {
+            return when (httpMethod) {
+                HttpMethod.Get -> toHttpGetRequest(responseAdapterCache)
+                HttpMethod.Post -> toHttpPostRequest(responseAdapterCache)
+            }
+        } catch (e: Exception) {
+            throw ApolloSerializationException(
+                message = "Failed to compose GraphQL network request",
+                cause = e
+            )
+        }
+    }
+
+    private fun <D : Operation.Data> ApolloRequest<D>.toHttpGetRequest(
+        responseAdapterCache: ResponseAdapterCache
+    ): HttpRequest {
+        val url = HttpUrlBuilder(serverUrl, mapOf(
+            "query" to operation.queryDocument(),
+            "operationName" to operation.name(),
+            "variables" to operation.variablesJson(responseAdapterCache)
+        )).build()
+
+        return HttpRequest(
+            method = HttpMethod.Get,
+            url =  url,
+            headers = executionContext[HttpRequestParameters]?.headers ?: emptyMap(),
+            body = null
+        )
+    }
+
+    private fun <D : Operation.Data> ApolloRequest<D>.toHttpPostRequest(
+        responseAdapterCache: ResponseAdapterCache
+    ): HttpRequest {
+        val requestBody = OperationRequestBodyComposer.compose(
+            operation = operation,
+            autoPersistQueries = false,
+            withQueryDocument = true,
+            responseAdapterCache = responseAdapterCache
+        )
+
+        return HttpRequest(
+            method = HttpMethod.Post,
+            url = serverUrl,
+            headers = executionContext[HttpRequestParameters]?.headers ?: emptyMap(),
+            body = object : HttpBody {
+                override val contentType = requestBody.contentType
+                override val contentLength = -1L
+
+                override fun writeTo(bufferedSink: BufferedSink) {
+                    requestBody.writeTo(bufferedSink)
+                }
+            }
+        )
+    }
+
+}
