+++ apollo-runtime-kotlin/src/commonMain/kotlin/com/apollographql/apollo3/network/http/HttpExecutionContext.kt
@@ -1,4 +1,4 @@
-package com.apollographql.apollo3.network
+package com.apollographql.apollo3.network.http
 
 import com.apollographql.apollo3.api.ExecutionContext
 import com.apollographql.apollo3.api.RequestContext
