+++ apollo-runtime-kotlin/src/commonMain/kotlin/com/apollographql/apollo3/network/http/HttpMethod.kt
@@ -1,0 +1,5 @@
+package com.apollographql.apollo3.network.http
+
+enum class HttpMethod {
+  Get, Post
+}
