+++ apollo-runtime-kotlin/src/commonMain/kotlin/com/apollographql/apollo3/network/http/HttpUrl.kt
@@ -1,0 +1,5 @@
+package com.apollographql.apollo3.network.http
+
+expect class HttpUrlBuilder(baseUrl: String, queryParameters: Map<String, String>) {
+  fun build(): String
+}
