+++ samples/multiplatform/kmp-lib-sample/src/commonMain/kotlin/com/apollographql/apollo3/kmpsample/data/ApolloCoroutinesRepository.kt
@@ -11,10 +11,8 @@ import com.apollographql.apollo3.kmpsample.fragment.RepositoryFragment
 import com.apollographql.apollo3.kmpsample.type.OrderDirection
 import com.apollographql.apollo3.kmpsample.type.PullRequestState
 import com.apollographql.apollo3.kmpsample.type.RepositoryOrderField
-import com.apollographql.apollo3.network.HttpResponseInfo
+import com.apollographql.apollo3.network.http.HttpResponseInfo
 import com.apollographql.apollo3.network.http.ApolloHttpNetworkTransport
-import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.flow.single
 
 /**
  * An implementation of a [GitHubDataSource] that shows how we can use coroutines to make our apollo requests.
